+++ data-raw/ex_binomial.R
@@ -1,8 +1,8 @@
 # Define the model
 segments = list(
   y | trials(N) ~ 1,  # constant rate
-  1 ~ 0 + x,  # joined changing rate
-  1 ~ 1 + x  # disjoined changing rate
+  ~ 0 + x,  # joined changing rate
+  ~ 1 + x  # disjoined changing rate
 )
 
 # Simulate data
@@ -10,7 +10,17 @@ empty = mcp::mcp(segments, family = binomial(), sample = FALSE)
 ex_binomial = tibble::tibble(
   x = 1:100,
   N = sample(10, length(x), replace=TRUE),
-  y = empty$func_y(x, N, 2, 0, -0.25, 0.05, 35, 70))
+  y = empty$simulate(
+    x,
+    N,
+    cp_1 = 35,
+    cp_2 = 70,
+    int_1 = 2,
+    int_3 = 0,
+    x_2 = -0.25,
+    x_3 = 0.05)
+  )
 
 # Save to mcp
+ex_binomial = data.frame(ex_binomial)
 usethis::use_data(ex_binomial, overwrite = TRUE)